Move FindBoolAttribute into extension class, move QKeyValue class into separate file
