+++ Source/Core/AST/ICarriesAttributes.cs
@@ -39,22 +39,20 @@ public interface ICarriesAttributes
     }
     return layers.Distinct().OrderBy(l => l).ToList();
   }
-
-  // Look for {:name string} in list of attributes.
-  public string FindStringAttribute(string name)
-  {
-    return QKeyValue.FindStringAttribute(Attributes, name);
-  }
-
-  public void AddStringAttribute(IToken tok, string name, string parameter)
-  {
-    Attributes = new QKeyValue(tok, name, new List<object>() {parameter}, Attributes);
-  }
-
-
 }
 
 public static class CarriesAttributesExtensions {
+  
+  // Look for {:name string} in list of attributes.
+  public static string FindStringAttribute(this ICarriesAttributes destination, string name)
+  {
+    return QKeyValue.FindStringAttribute(destination.Attributes, name);
+  }
+
+  public static void AddStringAttribute(this ICarriesAttributes destination, IToken tok, string name, string parameter)
+  {
+    destination.Attributes = new QKeyValue(tok, name, new List<object>() {parameter}, destination.Attributes);
+  }
   
   public static void CopyIdFrom(this ICarriesAttributes destination, IToken tok, ICarriesAttributes src)
   {
